requirements.txt
kivy==2.3.0
kivymd==1.2.0
google-generativeai==0.4.1
pillow==10.3.0
pyjnius==1.6.1
python-dotenv==1.0.1
